QuantumStatePlots

Documentation for QuantumStatePlots.

QuantumStatePlots.Contour

QuantumStatePlots.Heatmap

QuantumStatePlots.Surface

QuantumStatePlots.plot_all

QuantumStatePlots.plot_wigner

QuantumStatePlots.plot_wigner

QuantumStatePlots.plot_wigner

QuantumStatePlots.plot_ρ

QuantumStatePlots.Contour — Type

Contour <: PlotMethod

Contour representation for Wigner function of quantum state.

QuantumStatePlots.Heatmap — Type

Heatmap <: PlotMethod

Heatmap representation for Wigner function of quantum state.

QuantumStatePlots.Surface — Type

Surface <: PlotMethod

Surface representation for Wigner function of quantum state.

QuantumStatePlots.plot_all — Method

plot_all(
    ws::WignerSurface, state::AbstractState;
    state_n::Integer=0, size=(1100, 900), file_path=nothing
)

Plot Wifner function and density matrux of given quantum state in both Surface and Contour representation.

state_n=n: truncate to nth photon number

Examples

julia> using QuantumStateBase
julia> state = CoherentState(α(5., π/4));
julia> wf = WignerFunction(-10:0.1:10, -10:0.1:10);
julia> plot_all(wf(state), state)

QuantumStatePlots.plot_wigner — Method

plot_wigner(ws::WignerSurface, Contour; size=(1100, 900), file_path=nothing)

Plot Wigner function of given quantum state in Contour representation.

Examples

julia> using QuantumStateBase
julia> state = CoherentState(α(5., π/4));
julia> wf = WignerFunction(-10:0.1:10, -10:0.1:10);
julia> plot_wigner(wf(state), Contour)

QuantumStatePlots.plot_wigner — Method

plot_wigner(ws::WignerSurface, Heatmap; size=(1100, 900), file_path=nothing)

Plot Wigner function of given quantum state in Heatmap representation.

Examples

julia> using QuantumStateBase
julia> state = CoherentState(α(5., π/4));
julia> wf = WignerFunction(-10:0.1:10, -10:0.1:10);
julia> plot_wigner(wf(state), Surface)

QuantumStatePlots.plot_wigner — Method

plot_wigner(ws::WignerSurface, Surface; size=(1100, 900), file_path=nothing)

Plot Wigner function of given quantum state in Surface representation.

Examples

julia> using QuantumStateBase
julia> state = CoherentState(α(5., π/4));
julia> wf = WignerFunction(-10:0.1:10, -10:0.1:10);
julia> plot_wigner(wf(state), Surface)

QuantumStatePlots.plot_ρ — Method

plot_ρ(state::AbstractState; state_n::Integer=0, size=(1100, 900), file_path=nothing)

Plot density matrux of given quantum state.

state_n=n: truncate to nth photon number

Examples

julia> using QuantumStateBase
julia> state = CoherentState(α(5., π/4));
julia> plot_ρ(state)